chore: change test keys definition to use it on dev
