+++ decidim-bulletin_board-app/config/environments/development.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "test/private_keys"
+
 Rails.application.configure do
   # Settings specified here will take precedence over those in config/application.rb.
