+++ decidim-bulletin_board-app/lib/test/private_keys.rb
@@ -1,0 +1,83 @@
+# frozen_string_literal: true
+
+# rubocop:disable Metrics/LineLength
+
+module Test
+  module PrivateKeys
+    class << self
+      def authority_private_key_json
+        {
+          kty: "RSA",
+          n: "pNgMt8lnPDD3TlWYGhRiV1oZkPQmnLdiUzwyb_-35qKD9k-HU86xo0uSgoOUWkBtnvFscq8zNDPAGAlZVokaN_z9ksZblSce0LEl8lJa3ICgghg7e8vg_7Lz5dyHSQ3PCLgenyFGcL401aglDde1Xo4ujdz33Lklc4U9zoyoLUI2_viYmNOU6n5Mn0sJd30FeICMrLD2gX46pGe3MGug6groT9EvpKcdOoJHKoO5yGSVaeY5-Bo3gngvlgjlS2mfwjCtF4NYwIQSd2al-p4BKnuYAVKRSgr8rYnnjhWfJ4GsCaqiyXNi5NPYRV6gl_cx_1jUcA1rRJqQR32I8c8QbAXm5qNO4URcdaKys9tNcVgXBL1FsSdbrLVVFWen1tfWNfHm-8BjiWCWD79-uk5gI0SjC9tWvTzVvswWXI5weNqqVXqpDydr46AsHE2sG40HRCR3UF3LupT-HwXTcYcOZr5dJClJIsU3Hrvy4wLssub69YSNR1Jxn-KX2vUc06xY8CNIuSMpfufEq5cZopL6O2l1pRsW1FQnF3s078_Y9MaQ1gPyBo0IipLBVUj5IjEIfPuiEk4jxkiUYDeqzf7bAvSFckp94yLkRWTs_pEZs7b_ogwRG6WMHjtcaNYe4CufhIm9ekkKDeAWOPRTHfKNmohRBh09XuvSjqrx5Z7rqb8",
+          e: "AQAB",
+          kid: "b8dba1459df956d60107690c34fa490db681eac4f73ffaf6e4055728c02ddc8e",
+          d: "Uh3KIBe1VJez6pLbBUrYPlmE2N-3CGSWF46qNX62lq6ofB_b8xTJCuaPonJ3iYoE0aPEeVDrefq5m3-0wFXl-LQPgXlMj_1_7UgB9jeuSZ_N1WDK6P2EJPx5YS09O1gkpVxK7Mx_sZQe77wmUUH-eI7tg__qfUrB7E0Yn_cTpBATI2qlYaQsz6-A7e1MVvixq_ilmzVAZvuBrPp5mCZVb6FlXrV_PU9-UPIrD3O1La1lfO6SPBSbSGQkmGHwD2QbkHn9D_R_Vs-z_0TkM_dX71jIPQhrle3pN222KuJ8eQqwr9QP6biQMBuT5eKgr3MVtfUDRpp4sCEq9GIFwSd8LvbmGPrOoz8ueOEQ05nisIBQuOTYiWpYs2CEV062HR1bLFRLDUcSlflGNr0bgiXTUFx4wxRG06OaI-rQ6nG3M8TE0I0phMNCG3c7YyV28z_k2I65oQF9aKtiwFwc0YsUSGPTOFZGWHuCCPLm0lFeebpI_JIYqIv70NJxbSZEBY8DAIqZPqP6y_CRo2_C7piCgsjg9pnF8cp45vz4L6DWZ0Tumc_5aRuqIBkYXXwP9TjqhzxL-2SQHIqUAjj6Y6S35tZT6ekZSbnPIKX_e42y6bDT_Ztf01QfKiTkcx3_I8RwOuh6CzJzr72AykQpU3XKOKF1x1GBtYyrno4jG5LgaGE",
+          p: "1UARZ-rRnpKG5NHKlXTys3irCy-d91edHL3fEIzDKvhMRQCIWh7dt8l0_sIpcBF-EbVilbFKj7yfgZBTr8EkAXHgweayK8rnlMqi2jte1_u-5DBtrGVVUTSQltSLDOZHK5QfUxVK6Bbk8K5ROLvef91oNgnSNWNOeoCZdlS55nMZcAgY_6mxSuuMq54Tgy8o4Ip890-ZEYY6OSFXhU-ieoGO4Jw--c6QzmCa3gGo2oVClidMNaM1jquK4Pj6xaoxR2NWeIX9Ix7k1P2B24pegyHXjSIpQ6JYdn352VViXi2tx7TTJh6ClNVjgoRmL4Gfy_IJNx0GhF5OB3yughUc7w",
+          q: "xePJGBt466qM9F0BPxWFjyWbIs_GNXr-lBGASui0Z94cfgFbsZwqRsWQEf7jDVQsDNVnPSWZ_Wd6UqoQaIxc0tE8gaokPG6A4EUDyoLaZ231ZydDVoWof8FnPDaJwrcPwZ4R6ZLKGmkfytCZuU9I_9B4uuV0dyjEzKfS-Os3UcLumKPlgJ71OZAb49GTqUHuTePcSJjyYOYXx6eE7i_1m8TjU9Ut18BJNQhLqWmerA6X1ijbR2_syY6GXhGSfciSBH8xVkiUnqXb2jt1bE8nwWw-Sam5ikjzNbXqqs978IcCE5HTddQmy99bwuArA8PLqIFj3OOO1CSo8oyn2XDgMQ",
+          dp: "Diky_rOZN-6DBq7nxQT_GOvqb9O5qbMnu8DgDzlJvJDAf9SJOXLTRmEaY9CA7_A5bvOcmFQtn13nObNb20_4FCB7zGSFcGMI_dh2-Ab5RV5yTrTok4onID1dXKbAlRq1ny825U2Eq-TZTyJEQoA3RkZtpSkBzInLrFbd2f3GWodKKSZggpnCLDd4H-1fXlbDYCXSJpoikAdZ1nFgXnnrUDdKRaAajnwpIYtIvXVewSQYR-BULzunUtIRZt8hx_6FRzhRha9gH_TtPTeYZ_vISuz0Y2rhUpx1Q2kaLlR9M8PUxm47l0xvX3LMKN6h6oWxFtn7wq0qwZ-Bjv24mOrOAQ",
+          dq: "nXGD10hURrwk9W7hxP0sjB2Rdnr06iv3THs4JWFL16_h32bZO1BSWoho_chbgYlMmtFXGFFIWVLxAcAI2gWC_MA4cbmapvIMW2LNh1vgxJW5v95_NuGUlECeEEwcAu1-_b7z5XBCmAy3nLem9sbb_5wv0hMpPH0VRvbnZeBO3SBIkO0lddYCqU-8wN9HqkyoexQleSUnAm1O0iy4GIHT2aEmdNaRaKy2EhmNiTZdZeseZueOvyGPtTVONp2ofacMdcN0z39jr22qo9DWtdusd7nVPOpqkllEF6GrGUeHBnGD92n4YjDuxRnqefu8fXxUFrcLav0p8CNSv9ek291woQ",
+          qi: "w6hfKEBLLHRWPkjajgxZyyetj-UFfVkILRT0plOllJ2JV8whcOXRXbiXH2r8zqMeyMFrrMwmuvv4TVQaruKB0ZQOG7Tz5Lw0RZEREOLnBwc3vSi_iLd-jBz01LdExTpqsAHMkaMQR9x62J8DE1ZNxVdn3ELYKik0f1L2r_WErzhvT1uq69HAybUp6WHcFYH0PSqHg4LOneXAdU1_g-ji2Zn9dlA_2oYGQ5S6JXPV7v2IVbEFpxyVD1lPbFT0iKhyZZevictjgD_JGHveIVqsq5w0Csyz08h0oEW9hYEq-4bquMxSf18gjldoS5uQPD7FUECgL8bxsCdc4hP6UEKYGw"
+        }
+      end
+
+      def authority_private_key
+        Decidim::BulletinBoard::JwkUtils.import_private_key(authority_private_key_json)
+      end
+
+      def authority_public_key
+        authority_private_key.export
+      end
+
+      def trustees_private_keys_json
+        [
+          {
+            kty: "RSA",
+            n: "0TdbK3C79AMe_SnO5IibTJBUFrotTMHEI4YOH_L5VHHp59kO2CZSEgH9ZEnivOgUOz6W-v_BauVIa-ZLzQuiHh6gdA9j0pN5zEkPHYJqOIpML0JcnTEzyYEm5_oLBi_mWIPgKMa5nImVSwpwg9vf552JO_zkj3t3FzxsJ1TUy7sZnAXzLdzzGqgTvTFfqQWuItbNrVN8vk3SX01kwx-dnu1AokTMrnWGio0HhbzY9ZZMBmUxleGtcNwubJHi42k392LiJ0RVrAxvdyB3WVJxXXL7Tq8di0YEPnubLYYDJGyQ00qsGJ_UwI0H0oyxSnvdwYQoHElRHfYRseDLo-XEbuIhW-h_J6swWdYB8Mdpx12soEESVuBD4ep3PUJzChFd--bOUBCCNxOaQSY2p4pZ7Zf2bETSed0eRMLHfwPRZwxFFh7FdIqq0UzR_99jwjDoDneFuOYwwzec8t12RbImXcHsi2GPbLzlCzDAIkIydkUsw_J83fVXbPFWR86rpzWcC07uH-HXPoUn0DiI-0TwffAreDbumSaZUzJRLuJnJAPEg_XE1d6wpHuYwpTA5s5rXEioLYdogvjU30X7qr1i7_-aLci2jO9FdJGdYFTfeDqbUEwApxoEcScj3J5GwJdVJetmeW74vIZQkHm6LwAcQtRIyTGRh32zswNMv4jD-tc",
+            e: "AQAB",
+            kid: "11b75a77414348fe692ee9a98b896c250e992ee863726c27176913b806ccebf2",
+            d: "hxItCxuqTLxjGmEOkJ7dXI3lrBKgwmhqv-CBaAEPFYTB1YxeOBBtdfDQ7hmBiCJxRVqAUSYjcmb63aZ-E7BKBxYQ0z1IXiAWR1x3pL6Y403uJ6i5cEbev15hhuz4hxMPrN_fLZQ3KyfsOdlQkOC4aqlbe14nFbdJSLgavz3WcH2n6AIzYM3AKdqPe9rPfqkLJrT1bsxywc4uXpbx0cOsmHll_B1c49k-dl7cl2w3yj4xcQOOAs0Dk9SjKDb7e4RnvJCbn2xPv1b787YDN6ixRfrA4qjN8yfN_6T_p2FY0ABDdBZGi1NNIM8sWLO6lnLn1VMELlbrRuOhwux1rKffq-f0mUMs1uj8JpBYSso1INmWdOgB2H6qTQxzFEs6DS1R2e3Mm9sqHvg6ZT2uPe647seHhGDPmPceu3Qd8NPmc1wtOiuPaxVV77nn1ligMfvFY2ISenaOAkHncxgLavEJdhHRSE15ujjQjsUpHHVw4PhFnuVyjZ2zhqVJuK8kQVyq1hyAVhISiQR8JBUHjwne7oQn426nnnlGa_A3kbzChxjPsAIFd9bawBWvSGMt412VpBGdiJ-drTsvWBDk1AcRFw8Z0UIlxf4fKhZhW7oCGJ0LFAKkTZXOIb6LU9dATHgf5O3cHwWI0sU6mk4x_enDqi4xZpix5poGP5i6UOVXlsk",
+            p: "64cgMkAe9P-jheKebvLS2imlhjZncQymV1S15nnDyBErhvzBo5QFo55xgrvF4HOkreNqOURdbbJETrO-r5xJYzed_reajX_aVPhsHbF0DBs5PomLkTZh_K1M-MvHRKhxvRu5otdffYqq2K2MOrS0_RyYmNu4SkFNVugPxTPUMP496k31J06hTPRD3URXbO5_zKEZmQOenIWmZI4o9b8ujjehLykDnyJgKLsUQwyWCARx_R47aps5C7JqAXGsWEk30r7RyIeV4WWl1yOPm8vyqVKBS2f7adcXtcD1vpMNalzfm_73t5MaPUnZhMJ9CutcuDSXA1dhNN9hK__SGFr3vQ",
+            q: "42bBO_p4UCMgKAPPPFMYBoOKA3dCUPb9rJs_pEAJuzetpA7l009cwBnu8W5Pn9_ybTc1U588M6U5K87wpINjhumZo84KK4g-cWM9wL_WHYsctKmxvU8UvC5nNS0RgtRPNspIu_1yLwGO9t-BKHCa3D2mQh1hX1otOVGeX_i1J070UczPQwDtMCKaho4cLBjQCLLFVn-L3YyfN3dFnxaAS-aDFXklps2H1n87smgkrsA1q3unNpfVh2MRtNRolKmID52pMUCW0QxbCnEXBtyarsUh9JvKkzQMyS1_lVXPmRncPEQ4pbmA03bIno2twp-_ziPD2zlo256ybJ0apmhMIw",
+            dp: "tXZlgvHOKBrxpQyl3nttorJlmlM-PWZunmhgwVLdWb6I17bThjFcauR_Nidv3G-c2HCXfcnDZensIOrxpqF2FWBbdU2GYTC5SbAgtHvtGgq7fwhAJMBOnHv1CtcOD0LsZrP-wYzIQJRe7xQ_L1fc_gpTM6pNgcBwUt32kIxPFvQ7RdOMZx14vy6PPuIRrZbPlhwk-bRFNvjt6xlwS25h8ya_I_Oom6nDhn4krUf8K-8o18i8JHW1bOKoM7ad_2NiBBgKz0JPYtqK_IekdHec3YH4pd1WYgRbeyhQ7eSkeGdXzPuTZbeLht7oHRYvewM9a5rxIpznAvJ8DodKGoOzzQ",
+            dq: "IWaEV2TwqwSdwoLcu4WQ79P-P4gSy4WbuPCMpnlpdIA3FvPBOgpcGzMz724-b8Ma_oOvZ3rqAsPv48lfeyp5QeWp1W97-uzmDw1-koMa1fgK3VZXbvDgoMtICX13Jvo8MFYB3OrbDKBz-TrW47O2a0b61pcuKC08m0CrvHjxUra8GA-3GaZXMEzE2aKh7LAwTjbD2XKefoobELCKH1XUvlE7S7ohl6B0pULG5-t4B9rfhzKAWvpuOJaKkT7ppXxJKdIEeeF8xXPFI4F82lsotSbrV9xSItDwV9RA-xRBi8FKoed7UowuqeC2kuNIdF-2cna6W8bQ6gEjP-TLd0m2dQ",
+            qi: "yYF6EV4sZqxSxPbrsBXQ5Ekrbgc7q0VGH38etVabSU-VzslSWxKZIyCW9LPOvhe8xr8ooUmE3igza0hX0txi2DgaXdKSJXDSl3ImmVxA9b4GvYdl0M83PKpU_DwR0XRe4LSJjnRYVcquFrOIpfqsAII2odB-qAj1igo6O9DBZexZLptvGfhRCyZd5z86IKy8dibojNaET9irxUcj_Xzkb5EIg5_w00vLn9Xz967aV7_P_u_upVM5l0T0l1q_s9F4eGfbM-obHGTIizfw315B7ciu-j9NeufsKjDGOarou1lz066qRPsWadrifaA3PVgxIThzyEuzuyMRGA_ufMjYoQ"
+          }, {
+            kty: "RSA",
+            n: "sU2MnN5O0ATJua1eLX6Vy6mrRpcevfqqahWF2M2-uTgEJJxzb1wdQ03O9lb0Y3sG8HLMid9zNKxFN3fqQtkjm114H_K5BaxtJTHmo4qCnyx0FKVv0m52Pdnnomo3pvZaBetbt8YYjmC-LsOZUXotBIUQ3Bx5IokSJLT8tpMVuvTf9f0S2HLN7E73rHTvx-EgsRelazjKBnSZTV_lRbaS7StxOR_QQAlzi1OtRQ21Nn2JOjn4g-O3eCR3ROm1liE0zFeFc6i872zLRWMW1MS2-5ztcRnh0cLUBhYSseLO_qkM8SX4MXa7fvx2IquLTKj-5qnmUcMjyGVbMGXGSkZJJuYwubqmTfbI98PixL-Srj5qsvKRyFeWN0MPNA839dXQyxpCqBD2fOtrfmfZvYrVgVdPesCRq1XkrNRZSz1pR67ZjHmzYJllMdGX3BZFgdZIRXvMcF_UgMrt5Qerm_ezeHqUZRNwdFS3QQkhF1FynSqd2AEUuLz8JGHnOwEFDvj1sRn0asbu6iOhe85PP6FpxZgzSEZ9B26PZ3cCfVaXlZKnS00Vz1YOh8Hs9gQ2SZxbq6CuL6fII9KNnpvjlxo5DaelqzTU4nNWs9lRGxHceFTdzGJEgoUaEv_y-s-HxynD9_zQJN-hiEE6snq99RPglirjl-veaXBVfrdQ6JUrIQ8",
+            e: "AQAB",
+            kid: "f5f2e3d81612ea6ce47b9013f4cf804536ad736febaf62bea3f551b7acab34b0",
+            d: "g0rbvyXX0q5Y-sfon2PHZQMFM8agGypNmzTI4pyoiNt0JfA8gtABU1-AzGXjnu7xWoiH8einlO-kacjAjjPg7CGw3GeOnybpw8tJ8PyWcWwnoUpIndjycia0Ik7pWxe93fhE-LcWZAg7jMGG_RCt2VT4ASaMhQhfiRMEt-X_B2xhk_SNSf7pgf2P5gLJtId-_-R0uAtbfqBFDUjfrFngPwQWWai3O1REXX3fJu1ZrToUritTXcb4i1zlfQBbWQrJ4C9HQThIgDtd4snXlEL_BR9VrJTv5baR7tfVjChaA7YHQWNOgMJdYVFt4YqKIoBaEVDqns0T6Ec12Ybj6QhXDd0xW9n9Adilhdb1MpsLPvD2r3zUAvAP4m67Wr9cLYBqR-nDmVjsigT7WDGmneun_AKCnymoBjwmsqjRw4fve04C2kt8qGjjoJWxxk399N6pGmnsbYmAwc9lNumvjHfhGf_SFyOjdn3vvx0E_6Nlh2KP8b4Lf0pFNzuabcqhyGmgmlyjSsB-UGEmH4XDo50IlhGi1OC-cjPMfg3Q34-pKt5S_Fq-ewYNIqbvlOK5AnUdH5mbClwff9PqCbe8w6v54QBo9fczD6sD3NO60pSbiBwfNdH5yW4L3LPb-KQlD-W3krtczCmKscuvpYOer8Exdg-dN_QO6kkWf-8kRjnfdcE",
+            p: "32RLYBr7ja3BlNe5lAA4Z1_kPZ-O1MUvqnEooC1l6wQSebsFk9DxA1ONACBzQUxRb0brISZ3xEBoWJyqVr73xG8eXn8OAxR6wvTaJbetDKiIbYevsrEY7Y7hYVfeyIZn73Iyrpakq9H5mnYEdFwRcNknqtUw1ekKnnHaGpFt9hMbmkaQluSz2mzinxRIC2Q0PfM2Te7Wc9sBv7QArB5ZvoUFZFnSLe-lUoWqwqALG4NQhFzJTFfmDhJQL2w3vDKHTwwUdPaC074fcTSzzayPVkcKrxdkb2_hn5aSxL9diU0BJ4bewQZKGSKlhKXRrfnjpZI6IO70P7jI7HiTMdjnEQ",
+            q: "yy8BozhISMNF_EJ5_57w9Z5cHQa8RPz38m6f1txTpsX3pJ2FPbLRpVMnqNMb1VjpSpxnAMplcExWjDwtpFdvGsxKutAeGjlT_p6jyoeiOleZ7cUDk0nwRxbL81n74F5Vpx9qxoge14wXxeroyZXO9iEgulzf2aGKimJDyWNgTLMGGtI-Ob2MxMGDdlfLInQ-bU9H5SWMuOuxCuwhh9rHIBMmrpk0yPUMoq_JAIutXSf2uMt3Lbb4TDXbQljy7b4KOMUw1X24AyvJggNzWWyPLsO7esbugRRzRjbNjie7s_w9_-wF_mnO5HhjdJxmCJl8C2dgVtDJXqYi96vxFjDGHw",
+            dp: "0EAoQFyHG0nbB82fdAhMjJoJJSVkupdkWykIbWJE-7UdIBwoE-eZdrdh37dgmoPx-qGLOqkox2wfmlPaO-nlAltwXTlmr5pc4vftkXHMQduu6ySttxB_o54qH1wSFK-vnrm_46ANvRuO696PQoY9wAzAfPInh5KrUo_iehgPAXJIriaQTgrAW6RcE16N1wi_5uyfOEGp6_FDjRnZ8pqMoR5Sq31MgYRBg0-Wo78c-7YLzojNI31husZ5t-vTfX9BQoCJZYwU5B3-I8O0xszSBuVNYXKWL4iGNgbhpcacBQ4jQFzrtdgOJaZ11NXcRaqJ_OT3iZIZYnn53PB-j96bcQ",
+            dq: "hmwc6z0rJWh7fr8EM4eV5wnT6ygvmny6FVv2x5GQjj9e7TuBkSYi6Puwu1KfkXB3N3Nt-entXnRCthiCIb3utY3-ysUxhSv0U_nkZvOA8pRONC4phWJfYsjEOcWIjBIdHKn-XEZmETiP6dLWG7fO6qtUGbekKTlkjNUh_7SNW3D9UvVRUaYD4ZAdR0giDjNVsb_H0uQay6SyxoUeUeD96EyDRPP7JhBpQaCGnKPMGFt-MRf7SvVIJDNlUenDHmKiJJjhC_1WplJryGY8BCd8OvYnrLy7wg4jjYpoR-ollpgE9LHT7VNXG3RCaD2oi3vvwyOrawRKlSHY9H5-y62NBQ",
+            qi: "if4tAUfLhfWSWmWJ1o3o-0Ae-G8DFCTeO2fKGln-E-DF-CkoWuTTDcB3ATNjrdSqPm-zVOsb312PvVcRFUtLmhZ-YoEdrysQjlt-3yxXLk-vkQuDJCzD_v12WQOi7vb4hK8NU9g_N6wRUkrgrpDwGuhqjV0N9y1aHAcNBM4lBZub6jCJFirfw2WM-QBn0D0GGSH35GCw4aVqBKwLY_4qw7niMVzrpGpnbboIdPPKrNsiqub2sReB9413i2qspCRsf3MFNwW2TVc57mJRhDVh7f6z3LVsy6a359ZQnyKS-V1rqRzHq5kxkVgzkvsMpNhYdozlAvAtqv7dA99rro08Dw"
+          }, {
+            kty: "RSA",
+            n: "ySJSWG7P-nsN20gav_N4zSq7bzSXIFRawoYtxdfva782QmZ-PEbTZE0Ere31r9OHgNxkSpZjAAN4YTEvuAA7As2a6KrRuNFhu0AiBTk53dmqfEdM2PzQ7YMOUPFQiyPeQZ91Q5JG0Jly9b3wmGs63lm7iGU7p6Gpt6_aurbCf0P-nrDQ4fLpJDYPWfwg5i0CGd29uwzOTXjD1FT_zKyZi8mXJjjvME3FzReQe2_Z_ybP8ac2-bMT8wJanq2zeXbpTTYl3dkVQRbmcn7WSSk4uYsG_j-0_SGjvHgBDGiKWkjmB35mUwHRVhw5o2cznyjMJH7NszU7RFbt5lDMqxhZopm2uoWMt1HrVCOxOamihKi_Zv9NzJFdcHnCJjsajC85-TX1EozoQ3sPAE1aQhDr9Pobz6OboY2YWK0WlhdoxcBudsfOizmReYemUuLAnlkTIhe86el0bHPHqcE3ltVdw0U7EwgiyPZmYsXJFX8yRdP8WB92cNKPl-ynVys3tkmNdvJPl2rL4J-zvPmKIlfnMeJQDbn8L1HpI9pfU3tgWw00q6I8nPZ0mniOyKS-ZzaDuW75xe3BussuW-1L2-HrGuYj83P3a_sDUSP-_sdo3TC1JOvV3XaO1swOTkhZ_ICi5fNhWRARuvkJ8cmWlStz5RFtUshv8giZCTH0fYSWdwk",
+            e: "AQAB",
+            kid: "883403b5adb9e35048f53eaa8d9528d9f11d7b33560d8aeb9d11e40892b00312",
+            d: "vk9FTvt6umxYjhVJijobITW_p9dA74whXFGNWdeZaH_CHU3RP8NcHVlqhg_UGgux-NGP66sj_U1LR9s8ZYOCUoKsgl2ulos3R-_EwOCGyk08LN9h7SY5oaIRl7DdHupPUBRK5iav7oxi4_ZWgkFjsCODd-khatliNrWj5msq2TRC5XA0FLa8tlCYcfJIRsJrbIINFcFmKpOL54qNhuIBXuzdoYLZfwD3mbkkMkeBIxVcSDCJqfYxeAvQJrgAS3uiADat2Z2fqLi8T7WHmC3yyEytz0AERkbgIGIDy_PWhNrZHFoTA5OCuVF3kA8EHyKBFj8M7vX9IAqIixYEshcWJ0Bir7RIqCP4fWhx7CoUj8S2Xtc2iVw33jtTWR9-o3deGDz8VZVy6qBni0BTrUfiEgH6_5w0EA3Ywj1ilIs5hMVC0HAWLRAzGl53X5cjf5Ghy8YJgnQxDtLUnlJd3dKLDgOzDGYtmFq5anrd77OaB6EcXcPmRILs69RiIRVxppjIX4_lqxa6QRnx6yubddMhzmzJnjLLHrX_X7zca10rDNsFM0Hh8ogJz-3CdUx5j32FFyHG_auamzZY4YUpA0XHSwy9yVWpJFhzCvTE_3_dKXw_3B1ti1-zjMbNv-rW1Z48Dfn0FcKNkoz3z1J6GoxGjNXjaVe52XE5jVwlhAbiOaE",
+            p: "6d2o67BLj4rai3v6T8JYr34wxPTxBQroMvs47s953RmNttkIKF5ZnK59o11bnlTVXfYl7LSWddbJj_sl0T_AwoN-3vRHy6ilU3Cws8VIgrnLtglI9v90OsigvtDwsQiWDSInshF609BGk5PYWtPocYvr4kBpAbGfBYoCp8MNueS1Mmj4De_XTqU3zp0olcrInH2shfYktYDnqHGxRgh0jv1UK-d5jNmM8tBoDYhys9FRAGui8Q1SSnBt1D4pdQIQeJPAW9MVDA_JvNtelgMPiCFJ-wwnEUkqqmnrb0Vxc9OYmKC0HeXG0KXdbRUeyZI7aEi8xsjwuKPY4X1Hye7srQ",
+            q: "3CuaSYefXpmSpl79sYrr7_tafCXHcriw5yY5a-8cw_sEvsqV6Fr7Yk1n3xdjsAPMjiXWiQmnjMzYe4DGFC6VpYf_Be4mLJb8X-4SezH3mToc6D-DIh3l7zdkEZ4K6eVtfqNzSVgQLvsNTt2IkbjNz87PZsAjotrKUjiYl47MkSZQx6YbzF6RWo5WizrH-dhvdF-CCd4W04l6U3MvcyKR7WJ1CIvL0z2rV-iwTjnYHHWrMTUKYS4ErNDjWPEgfAJxeuk09FBjpGdddZ8lNrAARDYfkLKOqtTRQJHWqaiiUCctfeJ5LKYMvF_3a5n0t5C5xI3jItskPfys1_aaL-hDTQ",
+            dp: "aPOr0hGs5z82MrCxtcePCzrDYjBWVZ_-cpqhgzKjCuFj4rHlYlBtQM5h3cWQzy5dDqzfeinU-A8rJ61a3FwTAngGoMHZ0oDdqOOCuTqI_vmofQoPr6EWpSbSVK_pxFkJFd3FGLmVeKcYdTQ9PQ2qXAKiGkgzl6oBpne4F1eGumz0ECphX8fJxSaKM_P9ebOmcexIV1oOQVIXJFXZa4dx35W6h3V67AXZpkJ239PjPeVqx5iGaE1FJ9jqwPwgmL0TbTlJcyv4XxBaxaFhAktfVgUycTKT-cLvHcko-iec8ZD9A0njayNqJMq64zA-P9GbTYKPDVqyBPqq4w7Y_7t0hQ",
+            dq: "SdWthh8fSQBDd9u_RD0MZoxqidu5iTV6zem5tAKaKPcTXlBd-XAZgZvny57yZWLAxKLX0VkVgVho-PH0e2wr7zkuQvTchHOV6y3RsNLO7-nXK-OwvE4-f_yZaHcRveKc161EykHoG3lBu0L7u7K4AoqebxzcVo787pngQmM3Iy4wROPlyqGkKTQ2vbZWvejhZ6tn1ptOGF6wUS_9p6TePzeXe-g10H-L6I02IuWvHcwwPu5ErtZZ5wKeW4Nxm2TOXaO7iHs1DztC0mEuf6qCTF3CNCpNQejgy8CctoTA-hRJY0AB1wFQoBOmy8TzlmPIiKv0kgnfm5IS_NHMX2iKuQ",
+            qi: "gDTouz7wdR21v7ue1bViI6zgyinK8FGa5nGzvKBNBLn1r0oDg51RBcHsmsRM2tKZlNx4CM7mgM2OVHQjtqwv4iNcjXTT1xXgg-UKRVhg3Pl_xnAZRGLcF0R999hASBk13ROikhNy9WPRBeFhdfEO-byTks4ZQJzBbtaKvBfAEaeAYfTA3C8uKWV4tS0qSiSpVt11EWIABxXFyyC5Xv2YVfyVxNMZyxZVSsdSWxQhNTTnsV8T6MWFxXAp3HOd5xrOdTe67fac1LGr16SB0EiCXkV0ycIh_XLYkeoLTtgI6kc_MOJ1YDTaISKwXlq5HKD8HPnWsoA8m-dHN96vuUDReg"
+          }
+        ]
+      end
+
+      def trustees_private_keys
+        trustees_private_keys_json.map do |trustee_private_key_json|
+          Decidim::BulletinBoard::JwkUtils.import_private_key(trustee_private_key_json)
+        end
+      end
+
+      def trustees_public_keys
+        trustees_private_keys.map(&:export)
+      end
+    end
+  end
+end
+
+# rubocop:enable Metrics/LineLength
