+++ decidim-bulletin_board-app/spec/commands/vote_spec.rb
@@ -61,7 +61,7 @@ RSpec.describe Vote do
 
   context "when the client is a trustee" do
     let(:client) { Trustee.first }
-    let(:private_key) { DevPrivateKeys.trustees_private_keys.first }
+    let(:private_key) { Test::PrivateKeys.trustees_private_keys.first }
 
     it "broadcast invalid" do
       expect { subject }.to broadcast(:invalid, "Invalid client")
