+++ decidim-bulletin_board-app/spec/jobs/process_key_ceremony_step_job_spec.rb
@@ -8,7 +8,7 @@ RSpec.describe ProcessKeyCeremonyStepJob do
   let!(:pending_message) { create(:pending_message, election: election, message: message, client: trustee, private_key: private_key) }
   let(:election) { create(:election) }
   let(:trustee) { Trustee.first }
-  let(:private_key) { DevPrivateKeys.trustees_private_keys.first }
+  let(:private_key) { Test::PrivateKeys.trustees_private_keys.first }
   let(:message) { build(:key_ceremony_message, content_traits: content_traits, election: election) }
   let(:content_traits) { [] }
