+++ decidim-bulletin_board-app/spec/mutations/open_ballot_box_mutation_spec.rb
@@ -23,7 +23,7 @@ module Mutations
     let!(:election) { create(:election, status: :ready) }
     let(:authority) { Authority.first }
     let(:headers) { { "Authorization": authority.api_key } }
-    let(:signed_data) { JWT.encode(payload.as_json, DevPrivateKeys.authority_private_key.keypair, "RS256") }
+    let(:signed_data) { JWT.encode(payload.as_json, Test::PrivateKeys.authority_private_key.keypair, "RS256") }
     let(:payload) { build(:open_ballot_box_message, election: election) }
     let(:message_id) { payload["message_id"] }
