+++ decidim-bulletin_board-app/spec/mutations/vote_mutation_spec.rb
@@ -30,7 +30,7 @@ module Mutations
     let!(:election) { create(:election, status: :vote) }
     let(:authority) { Authority.first }
     let(:headers) { { "Authorization": authority.api_key } }
-    let(:signed_data) { JWT.encode(payload.as_json, DevPrivateKeys.authority_private_key.keypair, "RS256") }
+    let(:signed_data) { JWT.encode(payload.as_json, Test::PrivateKeys.authority_private_key.keypair, "RS256") }
     let(:payload) { build(:vote_message, election: election) }
     let(:message_id) { payload["message_id"] }
