+++ decidim-bulletin_board-app/spec/support/seeds.rb
@@ -1,5 +1,7 @@
 # frozen_string_literal: true
 
+require "test/private_keys"
+
 RSpec.configure do |config|
   config.before(:suite) do
     DatabaseCleaner.strategy = :truncation
